+++ actix/tests/test_actor.rs
@@ -252,51 +252,3 @@ fn test_run_interval() {
     // We wait 10 intervals by ~100ms
     assert_eq!(result.elapsed().as_secs(), 1);
 }
-
-mod future_cancellation_bug {
-    use futures_util::stream;
-
-    use super::*;
-
-    struct TestActor;
-
-    impl Actor for TestActor {
-        type Context = Context<Self>;
-
-        fn started(&mut self, ctx: &mut Self::Context) {
-            ctx.add_stream(Box::pin(stream::repeat(())));
-        }
-    }
-
-    impl StreamHandler<()> for TestActor {
-        fn handle(&mut self, _: (), _: &mut Self::Context) {}
-    }
-
-    #[derive(Message)]
-    #[rtype(result = "()")]
-    struct TestMsg;
-
-    impl Handler<TestMsg> for TestActor {
-        type Result = ();
-
-        fn handle(&mut self, _: TestMsg, ctx: &mut Self::Context) -> Self::Result {
-            let handle = ctx.run_later(Duration::from_millis(0), |_, _| {
-                System::current().stop_with_code(1);
-                panic!("This closure should never be called");
-            });
-            ctx.cancel_future(handle);
-        }
-    }
-
-    #[test]
-    fn not_present() {
-        let sys = System::new();
-        sys.block_on(async {
-            let addr = TestActor.start();
-            addr.do_send(TestMsg);
-            sleep(Duration::from_millis(10)).await;
-            System::current().stop();
-        });
-        sys.run().unwrap();
-    }
-}
